Add files via upload
Project Description
The Simple Music Player project is a basic web application that allows users to play, pause, and navigate through a list of audio tracks. The player features a clean and minimalistic design, making it easy to use and understand. It serves as an excellent project for beginners to learn about HTML, CSS, and JavaScript, and how these technologies can be combined to create functional web applications.